Branch new

+++ api/views/review.py
@@ -3,7 +3,6 @@ from ..serializers import ReviewSerializer
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from django.http import Http404
 
 class ReviewList(APIView):
     """
@@ -13,9 +12,9 @@ class ReviewList(APIView):
         reviews = Review.objects.all()
         serializer = ReviewSerializer(reviews, many=True)
         return Response(serializer.data)
-    
+
     def post(self, request, format=None):
-        serializer = ReviewSerializer(data=request.data)
+        serializer = ReviewSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -25,13 +24,21 @@ class ReviewDetails(APIView):
     def get(self, request, pk, format=None):
         try:
             review = Review.objects.get(pk=pk)
-            serializer = ReviewSerializer(review)
-            return Response(serializer.data)
         except Review.DoesNotExist:
-            raise Http404
+            return Response(status=status.HTTP_404_NOT_FOUND)
         
-    def put(self, request, pk, format=None):
-        review = Review.objects.get(pk=pk)
+        serializer = ReviewSerializer(review)
+        return Response(serializer.data)
+        
+    def put(self, request, pk, format=None):        
+        try:
+            review = Review.objects.get(pk=pk)
+        except Review.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        
+        if review.user.id != request.user.id:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+        
         serializer = ReviewSerializer(review, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -39,6 +46,14 @@ class ReviewDetails(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
+        try:
+            review = Review.objects.get(pk=pk)
+        except Review.DoesNotExist:
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        
+        if review.user.id != request.user.id:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+        
         review = Review.objects.get(pk=pk)
         review.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
